Rename provider to aquarius
