+++ .dockerignore
@@ -1,3 +1,3 @@
-provider.egg-info
+aquarius.egg-info
 .eggs
 .tox
