+++ aquarius/constants.py
@@ -4,6 +4,6 @@ class ConfigSections:
 
 
 class BaseURLs:
-    BASE_PROVIDER_URL = '/api/v1/provider'
+    BASE_AQUARIUS_URL = '/api/v1/aquarius'
     SWAGGER_URL = '/api/v1/docs'  # URL for exposing Swagger UI (without trailing '/')
-    ASSETS_URL = BASE_PROVIDER_URL + '/assets'
+    ASSETS_URL = BASE_AQUARIUS_URL + '/assets'
