+++ config.ini
@@ -49,5 +49,5 @@ db.collection = protokeeper
 ;db.app_key=d068996d8d5b1a66cfc61dc3a83fa7ee
 
 [resources]
-;; These consitute part of the provider url which is used in setting the `api_url` in the `OceanContractsWrapper`
-provider.url = http://localhost:5000
+;; These consitute part of the aquarius url which is used in setting the `api_url` in the `OceanContractsWrapper`
+aquarius.url = http://localhost:5000
