+++ config.ini.template
@@ -47,12 +47,12 @@ keeper.network = ${KEEPER_NETWORK}
 market.address = ${MARKET_ADDRESS}
 auth.address = ${AUTH_ADDRESS}
 token.address = ${TOKEN_ADDRESS}
-provider.address = ${PROVIDER_ADDRESS}
+aquarius.address = ${AQUARIUS_ADDRESS}
 
 [resources]
 azure.account.name = ${AZURE_ACCOUNT_NAME}
 azure.account.key = ${AZURE_ACCOUNT_KEY}
 azure.container = ${AZURE_CONTAINER}
 
-; These consitute part of the provider url which is used in setting the `api_url` in the `OceanContractsWrapper`
-provider.url = ${PROVIDER_URL}
+; These consitute part of the aquarius url which is used in setting the `api_url` in the `OceanContractsWrapper`
+aquarius.url = ${AQUARIUS_URL}
