+++ setup.py
@@ -28,19 +28,19 @@ setup(
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
     ],
-    description="🐳 Ocean Provider.",
+    description="🐳 Ocean aquarius.",
     install_requires=requirements,
     license="Apache Software License 2.0",
     long_description=readme,
     long_description_content_type="text/markdown",
     include_package_data=True,
-    keywords='ocean-provider',
-    name='ocean-provider',
-    packages=find_packages(include=['provider', 'provider.app']),
+    keywords='ocean-aquarius',
+    name='ocean-aquarius',
+    packages=find_packages(include=['aquarius', 'aquarius.app']),
     setup_requires=setup_requirements,
     test_suite='tests',
     tests_require=test_requirements,
-    url='https://github.com/oceanprotocol/provider',
+    url='https://github.com/oceanprotocol/aquarius',
     version='0.1.1',
     zip_safe=False,
 )
