+++ tox.ini
@@ -9,7 +9,7 @@ python =
 [testenv:flake8]
 basepython = python
 deps = flake8
-commands = flake8 provider
+commands = flake8 aquarius
 
 [testenv]
 passenv = CODACY_PROJECT_TOKEN
@@ -23,7 +23,7 @@ deps =
 ;     -r{toxinidir}/requirements.txt
 commands =
     pip install -U pip
-    coverage run --source provider -m py.test  --basetemp={envtmpdir}
+    coverage run --source aquarius -m py.test  --basetemp={envtmpdir}
     coverage report
     coverage xml
     python-codacy-coverage -r {toxinidir}/coverage.xml
